+++ lib/rrj/models/mongoid.rb
@@ -7,13 +7,17 @@ module RubyRabbitmqJanus
     # Store instance information for MongoID database
     class JanusInstance
       include Mongoid::Document
-      include RubyRabbitmqJanus::Models::JanusInstanceConcern
+      include RubyRabbitmqJanus::Models::JanusInstanceMethods
+      include RubyRabbitmqJanus::Models::JanusInstanceCallback
 
       field :instance, type: Integer
       field :session, type: Integer
       field :enable, type: Boolean
+      field :thread, type: Integer
 
-      set_callback(:destroy, :before) { destroy_before_action }
+      set_callback(:create, :after) { callback_create_after }
+      set_callback(:update, :after) { callback_update_after }
+      set_callback(:destroy, :before) { callback_destroy_before }
     end
   end
 end